+++ priv/templates/components/keyboard.exs
@@ -2,8 +2,9 @@
   keyboard: [
     name: "keyboard",
     args: [
-      variant: ["default", "outline", "transparent", "shadow", "unbordered"],
+      variant: ["default", "outline", "transparent", "shadow", "bordered", "gradient"],
       color: [
+        "natural",
         "white",
         "primary",
         "secondary",
@@ -12,13 +13,14 @@
         "warning",
         "danger",
         "info",
-        "light",
+        "silver",
         "misc",
         "dawn"
       ],
       size: ["extra_small", "small", "medium", "large", "extra_large"],
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "full", "none"],
       only: ["keyboard"],
+      helpers: [],
       module: ""
     ],
     optional: [],